C.2 Task 4 — review fixes: EOF breadcrumb, error-swallow comment, shared test_util
Addresses all four issues raised in PR #116 code review.

1. `TtyVetoUx::decide` now distinguishes EOF (`read_line` → `Ok(0)`)
   from operator-hit-Enter. On first EOF it emits one operator-facing
   breadcrumb to the writer (`(stdin closed; remaining vetoes default
   to KeepLocal)`), latched by an `eof_logged: bool` so N closed-stdin
   vetoes produce one notice rather than N. Then falls through to the
   existing empty-line branch — behavior (KeepLocal per spec §D4) is
   unchanged; this is purely a diagnostic improvement so an operator
   whose stderr is still readable can tell their session degraded into
   auto-default mode. Covered by new test
   `eof_emits_breadcrumb_once_and_defaults_remaining_to_keep_local`.

2. The `let _ = writeln!(...)` / `let _ = self.writer.flush()` pattern
   is now explicitly commented: prompt-write failures are swallowed
   because if stderr dies but stdin lives, the operator may type blind
   but the decision they enter still propagates correctly. Panicking on
   broken stderr would degrade a half-broken session into a hard crash.

3. `dummy_veto` is hoisted from two identical inline definitions in
   `interactive::tests` and `noninteractive::tests` to a single
   `pub fn` in `cli/src/veto/test_util.rs` (registered as
   `#[cfg(test)] pub(crate) mod test_util` in `mod.rs`). Both sibling
   test modules now import it via `super::super::test_util::dummy_veto`.
   Removes ~36 lines of duplicated fixture code with no semantic change.

4. The no-max-attempts re-prompt loop (pathological input could hang
   `decide` forever) is logged as issue #117 rather than fixed in-task:
   operator Ctrl-C and Task 5's planned `--non-interactive` ↔
   `--password-stdin` arg-pair validation already mitigate, and fixing
   it requires an additional design call (cap value, breadcrumb form)
   better scoped as its own slice.

Workspace 849 → 850 (one new test). Clippy + fmt + Python conformance +
spec freshness all clean. Existing 13 tests unchanged; behavior of
`scripted_eof_with_no_input_defaults_to_keep_local` (single-veto EOF)
still produces KeepLocal as before — the breadcrumb is new bytes on the
writer but the decision is identical.

NEXT_SESSION.md handoff updated to reflect: new test count (14), new
file (`test_util.rs`), new design decisions (EOF latch, swallow-comment
rationale, shared fixture), and issue #117 in the open-issues list.
